added expansiontile story + minor changes
