+++ lib/flutter_meragi_design.dart
@@ -68,4 +68,4 @@ export 'src/utils/platform.dart';
 export 'src/utils/scroll_handler.dart';
 export 'src/utils/shortcut_activators.dart';
 export 'src/utils/widget_state.dart';
-export 'src/components/expansion_tile.dart';
+export 'src/components/expansion_tile/expansion_tile.dart';
