+++ lib/src/theme/theme_tokens.dart
@@ -334,21 +334,15 @@ class ThemeToken extends ThemeExtension<ThemeToken> with _$ThemeTokenTailorMixin
 
   // ExpansionTile
   @override
-  final bool initiallyExpanded;
-  @override
   final Color expansionTileSplashColor;
   @override
   final Color expansionTileHoverColor;
   @override
   final InteractiveInkFeatureFactory expansionTileSplashFactory;
   @override
-  final bool shouldOnTileTapExpand;
-  @override
   final Color expansionTileBackgroundColor;
   @override
   final EdgeInsets expansiontitlePadding;
-  @override
-  final bool showExpandMoreIcon;
   @override
   final AnimationStyle expansionAnimationStyle;
   @override
@@ -528,14 +522,11 @@ class ThemeToken extends ThemeExtension<ThemeToken> with _$ThemeTokenTailorMixin
     required this.navigationRailBoxShadow,
 
     // ExpansionTile
-    required this.initiallyExpanded,
     required this.expansionTileSplashColor,
     required this.expansionTileHoverColor,
     required this.expansionTileSplashFactory,
-    required this.shouldOnTileTapExpand,
     required this.expansionTileBackgroundColor,
     required this.expansiontitlePadding,
-    required this.showExpandMoreIcon,
     required this.expansionAnimationStyle,
     required this.visualDensity,
     required this.expansionTileCollapsedShape,
