+++ lib/src/theme/theme_tokens.tailor.dart
@@ -147,14 +147,11 @@ mixin _$ThemeTokenTailorMixin on ThemeExtension<ThemeToken> {
   BorderRadius get navigationRailBorderRadius;
   Duration get navigationRailAnimationDuration;
   List<BoxShadow> get navigationRailBoxShadow;
-  bool get initiallyExpanded;
   Color get expansionTileSplashColor;
   Color get expansionTileHoverColor;
   InteractiveInkFeatureFactory get expansionTileSplashFactory;
-  bool get shouldOnTileTapExpand;
   Color get expansionTileBackgroundColor;
   EdgeInsets get expansiontitlePadding;
-  bool get showExpandMoreIcon;
   AnimationStyle get expansionAnimationStyle;
   VisualDensity get visualDensity;
   ShapeBorder get expansionTileCollapsedShape;
@@ -310,14 +307,11 @@ mixin _$ThemeTokenTailorMixin on ThemeExtension<ThemeToken> {
     BorderRadius? navigationRailBorderRadius,
     Duration? navigationRailAnimationDuration,
     List<BoxShadow>? navigationRailBoxShadow,
-    bool? initiallyExpanded,
     Color? expansionTileSplashColor,
     Color? expansionTileHoverColor,
     InteractiveInkFeatureFactory? expansionTileSplashFactory,
-    bool? shouldOnTileTapExpand,
     Color? expansionTileBackgroundColor,
     EdgeInsets? expansiontitlePadding,
-    bool? showExpandMoreIcon,
     AnimationStyle? expansionAnimationStyle,
     VisualDensity? visualDensity,
     ShapeBorder? expansionTileCollapsedShape,
@@ -528,20 +522,16 @@ mixin _$ThemeTokenTailorMixin on ThemeExtension<ThemeToken> {
           this.navigationRailAnimationDuration,
       navigationRailBoxShadow:
           navigationRailBoxShadow ?? this.navigationRailBoxShadow,
-      initiallyExpanded: initiallyExpanded ?? this.initiallyExpanded,
       expansionTileSplashColor:
           expansionTileSplashColor ?? this.expansionTileSplashColor,
       expansionTileHoverColor:
           expansionTileHoverColor ?? this.expansionTileHoverColor,
       expansionTileSplashFactory:
           expansionTileSplashFactory ?? this.expansionTileSplashFactory,
-      shouldOnTileTapExpand:
-          shouldOnTileTapExpand ?? this.shouldOnTileTapExpand,
       expansionTileBackgroundColor:
           expansionTileBackgroundColor ?? this.expansionTileBackgroundColor,
       expansiontitlePadding:
           expansiontitlePadding ?? this.expansiontitlePadding,
-      showExpandMoreIcon: showExpandMoreIcon ?? this.showExpandMoreIcon,
       expansionAnimationStyle:
           expansionAnimationStyle ?? this.expansionAnimationStyle,
       visualDensity: visualDensity ?? this.visualDensity,
@@ -805,7 +795,6 @@ mixin _$ThemeTokenTailorMixin on ThemeExtension<ThemeToken> {
           : other.navigationRailAnimationDuration,
       navigationRailBoxShadow:
           t < 0.5 ? navigationRailBoxShadow : other.navigationRailBoxShadow,
-      initiallyExpanded: t < 0.5 ? initiallyExpanded : other.initiallyExpanded,
       expansionTileSplashColor: Color.lerp(
           expansionTileSplashColor, other.expansionTileSplashColor, t)!,
       expansionTileHoverColor: Color.lerp(
@@ -813,14 +802,10 @@ mixin _$ThemeTokenTailorMixin on ThemeExtension<ThemeToken> {
       expansionTileSplashFactory: t < 0.5
           ? expansionTileSplashFactory
           : other.expansionTileSplashFactory,
-      shouldOnTileTapExpand:
-          t < 0.5 ? shouldOnTileTapExpand : other.shouldOnTileTapExpand,
       expansionTileBackgroundColor: Color.lerp(
           expansionTileBackgroundColor, other.expansionTileBackgroundColor, t)!,
       expansiontitlePadding:
           t < 0.5 ? expansiontitlePadding : other.expansiontitlePadding,
-      showExpandMoreIcon:
-          t < 0.5 ? showExpandMoreIcon : other.showExpandMoreIcon,
       expansionAnimationStyle:
           t < 0.5 ? expansionAnimationStyle : other.expansionAnimationStyle,
       visualDensity: t < 0.5 ? visualDensity : other.visualDensity,
@@ -1025,14 +1010,11 @@ mixin _$ThemeTokenTailorMixin on ThemeExtension<ThemeToken> {
             const DeepCollectionEquality().equals(navigationRailBorderRadius, other.navigationRailBorderRadius) &&
             const DeepCollectionEquality().equals(navigationRailAnimationDuration, other.navigationRailAnimationDuration) &&
             const DeepCollectionEquality().equals(navigationRailBoxShadow, other.navigationRailBoxShadow) &&
-            const DeepCollectionEquality().equals(initiallyExpanded, other.initiallyExpanded) &&
             const DeepCollectionEquality().equals(expansionTileSplashColor, other.expansionTileSplashColor) &&
             const DeepCollectionEquality().equals(expansionTileHoverColor, other.expansionTileHoverColor) &&
             const DeepCollectionEquality().equals(expansionTileSplashFactory, other.expansionTileSplashFactory) &&
-            const DeepCollectionEquality().equals(shouldOnTileTapExpand, other.shouldOnTileTapExpand) &&
             const DeepCollectionEquality().equals(expansionTileBackgroundColor, other.expansionTileBackgroundColor) &&
             const DeepCollectionEquality().equals(expansiontitlePadding, other.expansiontitlePadding) &&
-            const DeepCollectionEquality().equals(showExpandMoreIcon, other.showExpandMoreIcon) &&
             const DeepCollectionEquality().equals(expansionAnimationStyle, other.expansionAnimationStyle) &&
             const DeepCollectionEquality().equals(visualDensity, other.visualDensity) &&
             const DeepCollectionEquality().equals(expansionTileCollapsedShape, other.expansionTileCollapsedShape) &&
@@ -1195,14 +1177,11 @@ mixin _$ThemeTokenTailorMixin on ThemeExtension<ThemeToken> {
       const DeepCollectionEquality().hash(navigationRailBorderRadius),
       const DeepCollectionEquality().hash(navigationRailAnimationDuration),
       const DeepCollectionEquality().hash(navigationRailBoxShadow),
-      const DeepCollectionEquality().hash(initiallyExpanded),
       const DeepCollectionEquality().hash(expansionTileSplashColor),
       const DeepCollectionEquality().hash(expansionTileHoverColor),
       const DeepCollectionEquality().hash(expansionTileSplashFactory),
-      const DeepCollectionEquality().hash(shouldOnTileTapExpand),
       const DeepCollectionEquality().hash(expansionTileBackgroundColor),
       const DeepCollectionEquality().hash(expansiontitlePadding),
-      const DeepCollectionEquality().hash(showExpandMoreIcon),
       const DeepCollectionEquality().hash(expansionAnimationStyle),
       const DeepCollectionEquality().hash(visualDensity),
       const DeepCollectionEquality().hash(expansionTileCollapsedShape),
@@ -1379,16 +1358,13 @@ extension ThemeTokenBuildContextProps on BuildContext {
       themeToken.navigationRailAnimationDuration;
   List<BoxShadow> get navigationRailBoxShadow =>
       themeToken.navigationRailBoxShadow;
-  bool get initiallyExpanded => themeToken.initiallyExpanded;
   Color get expansionTileSplashColor => themeToken.expansionTileSplashColor;
   Color get expansionTileHoverColor => themeToken.expansionTileHoverColor;
   InteractiveInkFeatureFactory get expansionTileSplashFactory =>
       themeToken.expansionTileSplashFactory;
-  bool get shouldOnTileTapExpand => themeToken.shouldOnTileTapExpand;
   Color get expansionTileBackgroundColor =>
       themeToken.expansionTileBackgroundColor;
   EdgeInsets get expansiontitlePadding => themeToken.expansiontitlePadding;
-  bool get showExpandMoreIcon => themeToken.showExpandMoreIcon;
   AnimationStyle get expansionAnimationStyle =>
       themeToken.expansionAnimationStyle;
   VisualDensity get visualDensity => themeToken.visualDensity;
